backend/requirements.txt
flask == 2.3.2
pytest == 7.4.0
gunicorn == 20.1.0
flask-cors == 4.0.0
requests == 2.31.0
flask-rq2 == 18.3
flask-jwt-extended == 4.6.0
scikit-learn == 1.1.3
concrete-ml == 1.5.0
